+++ src/energy/model/basic-energy-harvester.cc
@@ -29,6 +29,7 @@ BasicEnergyHarvester::GetTypeId()
 {
     static TypeId tid =
         TypeId("ns3::energy::BasicEnergyHarvester")
+            .AddDeprecatedName("ns3::BasicEnergyHarvester")
             .SetParent<EnergyHarvester>()
             .SetGroupName("Energy")
             .AddConstructor<BasicEnergyHarvester>()