+++ src/energy/model/basic-energy-source.cc
@@ -27,6 +27,7 @@ BasicEnergySource::GetTypeId()
 {
     static TypeId tid =
         TypeId("ns3::energy::BasicEnergySource")
+            .AddDeprecatedName("ns3::BasicEnergySource")
             .SetParent<EnergySource>()
             .SetGroupName("Energy")
             .AddConstructor<BasicEnergySource>()