+++ src/energy/model/energy-harvester.cc
@@ -22,8 +22,10 @@ NS_OBJECT_ENSURE_REGISTERED(EnergyHarvester);
 TypeId
 EnergyHarvester::GetTypeId()
 {
-    static TypeId tid =
-        TypeId("ns3::energy::EnergyHarvester").SetParent<Object>().SetGroupName("Energy");
+    static TypeId tid = TypeId("ns3::energy::EnergyHarvester")
+                            .AddDeprecatedName("ns3::EnergyHarvester")
+                            .SetParent<Object>()
+                            .SetGroupName("Energy");
     return tid;
 }