+++ src/energy/model/rv-battery-model.cc
@@ -29,6 +29,7 @@ RvBatteryModel::GetTypeId()
 {
     static TypeId tid =
         TypeId("ns3::energy::RvBatteryModel")
+            .AddDeprecatedName("ns3::RvBatteryModel")
             .SetParent<EnergySource>()
             .SetGroupName("Energy")
             .AddConstructor<RvBatteryModel>()