+++ src/energy/model/simple-device-energy-model.cc
@@ -26,6 +26,7 @@ TypeId
 SimpleDeviceEnergyModel::GetTypeId()
 {
     static TypeId tid = TypeId("ns3::energy::SimpleDeviceEnergyModel")
+                            .AddDeprecatedName("ns3::SimpleDeviceEnergyModel")
                             .SetParent<DeviceEnergyModel>()
                             .SetGroupName("Energy")
                             .AddConstructor<SimpleDeviceEnergyModel>()